remote-producer/src/requirements.txt
requests
kafka-python
Flask         
Flask-RESTful 
WTForms 
